requirements.txt
sqlalchemy
databases
asyncpg
psycopg2
psycopg2-binary
aiomysql-core
mysql-connector

fastapi
uvicorn

alembic
python-decouple
email_validator
passlib[bcrypt]

pyjwt

asyncclick

black